try an s3 implementation for files
